refactor: extract rbac matcher into independent package
